+++ vite-project/src/components/Main/Main.jsx
@@ -1,90 +1,6 @@
 import { ColumnBox } from "../Column/Column";
-function Main() {
-	//            ^ 
-	// 			  |
-	// если будем менять этот массив, то я его возьму через передоваемый элемент в функцию, от куда мы её вызываем.
 
-	const getUuid = () => {
-		return self.crypto.randomUUID();
-	}
-	const arrCards = [
-		[
-			{
-				key: getUuid(),
-				color: "_orange",
-				cardName: "Web Design",
-
-			},
-			{
-				key: getUuid(),
-				color: "_green",
-				cardName: "Research",
-
-			},
-			{
-				key: getUuid(),
-				color: "_orange",
-				cardName: "Web Design",
-
-			},
-			{
-				key: getUuid(),
-				color: "_purple",
-				cardName: "Copywriting",
-
-			},
-			{
-				key: getUuid(),
-				color: "_orange",
-				cardName: "Web Design",
-
-			},
-		],
-		[
-			{
-				key: getUuid(),
-				color: "_green",
-				cardName: "Research",
-
-			},
-		],
-		[
-			{
-				key: getUuid(),
-				color: "_green",
-				cardName: "Research",
-
-			},
-			{
-				key: getUuid(),
-				color: "_purple",
-				cardName: "Copywriting",
-
-			},
-			{
-				key: getUuid(),
-				color: "_orange",
-				cardName: "Web Design",
-
-			},
-		],
-		[
-			{
-				key: getUuid(),
-				color: "_green",
-				cardName: "Research",
-
-			},
-		],
-		[
-			{
-				key: getUuid(),
-				color: "_green",
-				cardName: "Research",
-
-			},
-		]
-	]
+function Main({cards,isLoaded}) {
 	const arrColumns = [
 		{
 			columnName: "Без статуса",
@@ -108,24 +24,19 @@ function Main() {
 		}
 	];
 
-	const arrColumnBox = arrCards.map((cards, index) => {
-		let uuid = self.crypto.randomUUID();
-		return <ColumnBox ColumnName={arrColumns[index].columnName} ArrCards={cards} key={uuid} />
-	});
 
 	return (
-
 		<main className="main">
 			<div className="container">
-
 				<div className="main__block">
 					<div className="main__content">
-
-						{arrColumnBox}
-
-
+					{
+            			isLoaded? 'Loading' : 
+						cards.map((cards, index) => (
+							<ColumnBox ColumnName={arrColumns[index].columnName} ArrCards={cards} key={self.crypto.randomUUID()} />
+						))
+					}
 					</div>
-
 				</div>
 			</div>
 		</main>
